+++ src/Headers/FACG.h
@@ -32,7 +32,7 @@ void
 query_probes_Landau( void ) ;
 
 /**
-   @fn int FACG( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const int max_iters )
+   @fn size_t FACG( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const size_t max_iters )
    @brief Fourier Accelerated CG Landau gauge fixing
    @param lat :: the gauge links (overwritten)
    @param gauge :: gauge transformation matrices (overwritten)
@@ -46,20 +46,20 @@ query_probes_Landau( void ) ;
    @param max_iters :: the maximum number of iterations we will perform
    @warning is in general, faster than the steepest descent method at the cost of a little extra storage
  */
-int 
-FACG( struct site *__restrict lat , 
-      GLU_complex *__restrict *__restrict gauge , 
+size_t
+FACG( struct site *__restrict lat ,
+      GLU_complex *__restrict *__restrict gauge ,
       GLU_complex *__restrict *__restrict out ,
       GLU_complex *__restrict *__restrict in ,
-      const void *__restrict forward , 
-      const void *__restrict backward , 
-      const GLU_real *__restrict p_sq , 
+      const void *__restrict forward ,
+      const void *__restrict backward ,
+      const GLU_real *__restrict p_sq ,
       double *th ,
       const double acc ,
-      const int max_iters ) ;
+      const size_t max_iters ) ;
 
 /**
-   @fn int FASD( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const int max_iters )
+   @fn int FASD( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const size_t max_iters )
    @brief Fourier Accelerated CG Landau gauge fixing
    @param lat :: the gauge links (overwritten)
    @param gauge :: gauge transformation matrices (overwritten)
@@ -73,21 +73,21 @@ FACG( struct site *__restrict lat ,
    @param max_iters :: the maximum number of iterations we will perform
    @warning probably deprecated for the CG
  */
-int 
-FASD( struct site *__restrict lat , 
-      GLU_complex *__restrict *__restrict gauge , 
+size_t
+FASD( struct site *__restrict lat ,
+      GLU_complex *__restrict *__restrict gauge ,
       GLU_complex *__restrict *__restrict out ,
       GLU_complex *__restrict *__restrict in ,
-      const void *__restrict forward , 
-      const void *__restrict backward , 
-      const GLU_real *__restrict p_sq , 
+      const void *__restrict forward ,
+      const void *__restrict backward ,
+      const GLU_real *__restrict p_sq ,
       double *th ,
       const double acc ,
-      const int max_iters ) ;
+      const size_t max_iters ) ;
 
 
 /**
-   @fn int FASD_SMEAR( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const int max_iters )
+   @fn size_t FASD_SMEAR( struct site *__restrict lat , GLU_complex *__restrict *__restrict gauge , GLU_complex *__restrict *__restrict out , GLU_complex *__restrict *__restrict in , const void *__restrict forward , const void *__restrict backward , const GLU_real *__restrict p_sq , double *th , const double acc , const size_t max_iters )
    @brief Fourier Accelerated CG Landau gauge fixing
    @param lat :: the gauge links (overwritten)
    @param gauge :: gauge transformation matrices (overwritten by the product at each step)
@@ -100,16 +100,16 @@ FASD( struct site *__restrict lat ,
    @param acc :: the targeted gauge fixing accuracy
    @param max_iters :: the maximum number of iterations we will perform
  */
-int 
-FASD_SMEAR( struct site *__restrict lat , 
-	    GLU_complex *__restrict *__restrict gauge , 
+size_t 
+FASD_SMEAR( struct site *__restrict lat ,
+	    GLU_complex *__restrict *__restrict gauge ,
 	    GLU_complex *__restrict *__restrict out ,
 	    GLU_complex *__restrict *__restrict in ,
-	    const void *__restrict forward , 
-	    const void *__restrict backward , 
-	    const GLU_real *__restrict p_sq , 
+	    const void *__restrict forward ,
+	    const void *__restrict backward ,
+	    const GLU_real *__restrict p_sq ,
 	    double *th ,
 	    const double acc ,
-	    const int max_iters ) ;
+	    const size_t max_iters ) ;
 
 #endif
