+++ src/Headers/staples.h
@@ -25,7 +25,7 @@
 #define GLU_STAPLES_H
 
 /**
-   @fn void all_staples( GLU_complex stap[ NCNC ] , const struct site *__restrict lat , const int i , const int mu , const int dir , const int type )
+   @fn void all_staples( GLU_complex stap[ NCNC ] , const struct site *__restrict lat , const size_t i , const size_t mu , const size_t dir , const size_t type )
    @brief the computation of the basic unimproved staple.
    @param stap :: sum of the contributing mu-nu staples
    @param lat :: lattice gauge field
@@ -37,13 +37,13 @@
 void
 all_staples( GLU_complex stap[ NCNC ] , 
 	     const struct site *__restrict lat , 
-	     const int i , 
-	     const int mu , 
-	     const int dir , 
-	     const int type ) ;
+	     const size_t i , 
+	     const size_t mu , 
+	     const size_t dir , 
+	     const size_t type ) ;
 
 /**
-   @fn void all_staples_improve( GLU_complex stap[ NCNC ] , const struct site *__restrict lat , const int i , const int mu , const int dir , const int type )
+   @fn void all_staples_improve( GLU_complex stap[ NCNC ] , const struct site *__restrict lat , const size_t i , const size_t mu , const size_t dir , const size_t type )
    @brief he computation of the (over)improved staple(s)
    @param stap :: sum of the contributing mu-nu staples
    @param lat :: lattice gauge field
@@ -62,9 +62,9 @@ all_staples( GLU_complex stap[ NCNC ] ,
 void
 all_staples_improve( GLU_complex stap[ NCNC ] , 
 		     const struct site *__restrict lat , 
-		     const int i , 
-		     const int mu , 
-		     const int dir , 
-		     const int type ) ;
+		     const size_t i , 
+		     const size_t mu , 
+		     const size_t dir , 
+		     const size_t type ) ;
 
 #endif
